requirements.txt
discord
python-dotenv
pyyaml
boto3
